refactor(패키지 변경) : MessageType 패키지 이동
MessageType을 Enum이 상속받고있는 것이 공통으로 바뀌어서 패키지를 이동
